\name{StanBayesianErrorsMonoModels-package}
\alias{StanBayesianErrorsMonoModels-package}
\alias{StanBayesianErrorsMonoModels}
\docType{package}
\title{
\packageTitle{StanBayesianErrorsMonoModels}
}
\description{
\packageDescription{StanBayesianErrorsMonoModels}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{StanBayesianErrorsMonoModels}
\packageIndices{StanBayesianErrorsMonoModels}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{StanBayesianErrorsMonoModels}

Maintainer: \packageMaintainer{StanBayesianErrorsMonoModels}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
